ShapeManager. Правим баги когда после скейлинга по диагонали текст с длинными словами мог вылезать за пределы шейпа
