+++ e2e/fixtures/data/shape-fixed-width-long-word.data.ts
@@ -1,0 +1,19 @@
+import { SHAPE_PROPORTIONAL_SCALING_CORNERS } from './shape-scaling.data'
+
+export const SHAPE_FIXED_WIDTH_LONG_WORD = 'TESTTESTTEST'
+
+export const SHAPE_INITIAL_DIAGONAL_SCALE_TEXT = 'TEST'
+
+export const SHAPE_FIXED_WIDTH_LONG_WORD_FONT_SIZE = 48
+
+export const SHAPE_FIXED_WIDTH_TOLERANCE = 2
+
+export const SHAPE_INITIAL_EXPLICIT_SIZE_SCALE = 1.08
+
+export const SHAPE_NARROW_DIAGONAL_SCALE_X = 0.42
+
+export const SHAPE_NARROW_DIAGONAL_SCALE_Y = 0.72
+
+export const SHAPE_SMALL_PROPORTIONAL_GROW_SCALE = 1.15
+
+export const SHAPE_FIXED_WIDTH_LONG_WORD_CORNERS = SHAPE_PROPORTIONAL_SCALING_CORNERS
