+++ specs/test-utils/shape-helpers.ts
@@ -1,7 +1,6 @@
 import { Group, Point } from 'fabric'
 import type {
   ShapeNode,
-  ShapeTextWrapPolicy,
   ShapeTextNode
 } from '../../src/editor/shape-manager/types'
 import { BackgroundTextbox } from '../../src/editor/text-manager/background-textbox'
@@ -64,7 +63,6 @@ type MockShapeGroup = Group & {
   shapeManualBaseHeight: number
   shapeReplaceBoxWidth: number
   shapeReplaceBoxHeight: number
-  shapeTextWrapPolicy?: ShapeTextWrapPolicy
   shapeAlignHorizontal: 'left' | 'center' | 'right' | 'justify'
   shapeAlignVertical: 'top' | 'middle' | 'bottom'
   shapePaddingTop: number
