+++ src/editor/constants.ts
@@ -69,7 +69,6 @@ export const CLIPBOARD_CLONE_OBJECT_KEYS = [
   'shapeReplaceBoxWidth',
   'shapeReplaceBoxHeight',
   'shapeTextAutoExpand',
-  'shapeTextWrapPolicy',
   'shapeAlignHorizontal',
   'shapeAlignVertical',
   'shapePaddingTop',
