+++ src/editor/history-manager/constants.ts
@@ -50,7 +50,6 @@ export const OBJECT_SERIALIZATION_PROPS = [
   'shapeReplaceBoxWidth',
   'shapeReplaceBoxHeight',
   'shapeTextAutoExpand',
-  'shapeTextWrapPolicy',
   'shapeAlignHorizontal',
   'shapeAlignVertical',
   'shapePaddingTop',
